refactor: change the name of typesAndInterfaces to types and change all imports in codebase
