+++ app/components/ExamPages/ChordsIdentify.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { chordTextInstructions } from "@/app/lib/data/instructions";
-import { InputData, UserDataProps } from "@/app/lib/typesAndInterfaces";
+import { InputData, UserDataProps } from "@/app/lib/types";
 import { Box, Container, Stack, Typography } from "@mui/material";
 import { useRef } from "react";
 import {
