+++ app/components/ExamPages/KeySigIdentify.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { keySigInputInstructions } from "@/app/lib/data/instructions";
-import { InputData, UserDataProps } from "@/app/lib/typesAndInterfaces";
+import { InputData, UserDataProps } from "@/app/lib/types";
 import { Box, Container, Stack, Typography } from "@mui/material";
 import { useRef } from "react";
 import CardFooter from "../CardFooter";
