+++ app/components/ExamPages/TriadsNotateTemplate.tsx
@@ -1,11 +1,7 @@
 "use client";
 import { chordsNotationInstructions } from "@/app/lib/data/instructions";
 import triadsText from "@/app/lib/data/triadsText";
-import {
-  FormEvent,
-  InputState,
-  UserDataProps,
-} from "@/app/lib/typesAndInterfaces";
+import { FormEvent, InputState, UserDataProps } from "@/app/lib/types";
 import { Box, Container, Stack, Typography } from "@mui/material";
 import { useCallback, useEffect, useRef, useState } from "react";
 import CardFooter from "../CardFooter";
