+++ app/components/IdentifyKeySigs.tsx
@@ -3,12 +3,7 @@ import { ForwardedRef, forwardRef, useEffect, useMemo, useState } from "react";
 import { useClef } from "../context/ClefContext";
 import createInitialState from "../lib/createInitialState";
 import isCurrentDataFilled from "../lib/isCurrentDataFilled";
-import {
-  ChangeEvent,
-  Chord,
-  FormEvent,
-  InputData,
-} from "../lib/typesAndInterfaces";
+import { ChangeEvent, Chord, FormEvent, InputData } from "../lib/types";
 import FormInput from "./FormInput";
 import KeySigStaff from "./KeySigStaff";
 interface TextInput {
