+++ app/lib/createBlankStaves.ts
@@ -1,5 +1,5 @@
 import VexFlow from "vexflow";
-import { BlankStaves, CreateBlankStavesParams } from "./typesAndInterfaces";
+import { BlankStaves, CreateBlankStavesParams } from "./types";
 const VF = VexFlow.Flow;
 const { Stave } = VF;
