+++ app/lib/modifyScales.ts
@@ -12,7 +12,7 @@ import {
   ScaleData,
   StateInteraction,
   StaveNoteType,
-} from "./typesAndInterfaces";
+} from "./types";
 
 const { Accidental, StaveNote } = VexFlow.Flow;
