+++ app/lib/savePDF.ts
@@ -1,7 +1,7 @@
 import { getDownloadURL, ref, uploadBytes } from "firebase/storage";
 import html2canvas from "html2canvas";
 import jsPDF from "jspdf";
-import { InputState } from "./typesAndInterfaces";
+import { InputState } from "./types";
 import { storage } from "../../firebase/config";
 
 // Get a reference to the storage service, which is used to create references in your storage bucket
