+++ app/lib/setupRendererAndDrawNotes.ts
@@ -1,6 +1,6 @@
 import VexFlow from "vexflow";
 import createBlankStaves from "./createBlankStaves";
-import { RenderStavesAndNotesParams, BlankStaves } from "./typesAndInterfaces";
+import { RenderStavesAndNotesParams, BlankStaves } from "./types";
 const { Formatter, TickContext } = VexFlow.Flow;
 
 export const setupRendererAndDrawNotes = (
